Moved db related code to separate file
